+++ packages/eslint-plugin/src/rules/no-shadow.ts
@@ -1,4 +1,5 @@
 import {
+  ASTUtils,
   AST_NODE_TYPES,
   TSESLint,
   TSESTree,
@@ -341,31 +342,6 @@ export default util.createRule<Options, MessageIds>({
     }
 
     /**
-     * Finds the variable by a given name in a given scope and its upper scopes.
-     * @param initScope A scope to start find.
-     * @param name A variable name to find.
-     * @returns A found variable or `null`.
-     */
-    function getVariableByName(
-      initScope: TSESLint.Scope.Scope | null,
-      name: string,
-    ): TSESLint.Scope.Variable | null {
-      let scope = initScope;
-
-      while (scope) {
-        const variable = scope.set.get(name);
-
-        if (variable) {
-          return variable;
-        }
-
-        scope = scope.upper;
-      }
-
-      return null;
-    }
-
-    /**
      * Checks the current context for shadowed variables.
      * @param {Scope} scope Fixme
      */
@@ -404,7 +380,9 @@ export default util.createRule<Options, MessageIds>({
         }
 
         // Gets shadowed variable.
-        const shadowed = getVariableByName(scope.upper, variable.name);
+        const shadowed = scope.upper
+          ? ASTUtils.findVariable(scope.upper, variable.name)
+          : null;
         if (!shadowed) {
           continue;
         }